+++ src/main/scala/com/github/gchudnov/mtg/diagram/internal/BasicRenderer.scala
@@ -1,9 +1,9 @@
-package com.github.gchudnov.mtg.internal.diagram
+package com.github.gchudnov.mtg.diagram.internal
 
 import com.github.gchudnov.mtg.Diagram
 import com.github.gchudnov.mtg.Diagram.Theme
 import com.github.gchudnov.mtg.Diagram.Renderer
-import com.github.gchudnov.mtg.Diagram.Span
+import com.github.gchudnov.mtg.diagram.Span
 import com.github.gchudnov.mtg.Diagram.Tick
 import com.github.gchudnov.mtg.Diagram.Label
 import scala.collection.mutable.ListBuffer
