+++ src/main/scala/com/github/gchudnov/mtg/diagram/internal/BasicTranslator.scala
@@ -1,4 +1,4 @@
-package com.github.gchudnov.mtg.internal.diagram
+package com.github.gchudnov.mtg.diagram.internal
 
 import com.github.gchudnov.mtg.Domain
 import com.github.gchudnov.mtg.Value
@@ -8,7 +8,7 @@ import com.github.gchudnov.mtg.Diagram.Canvas
 import com.github.gchudnov.mtg.Diagram
 import com.github.gchudnov.mtg.Diagram.Theme
 import com.github.gchudnov.mtg.Diagram.Translator
-import com.github.gchudnov.mtg.Diagram.Span
+import com.github.gchudnov.mtg.diagram.Span
 import com.github.gchudnov.mtg.Diagram.View
 import com.github.gchudnov.mtg.Diagram.Tick
 import com.github.gchudnov.mtg.Diagram.Label
